requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
frozenlist==1.3.1
greenlet==1.1.3
idna==3.4
importlib-metadata==4.12.0
ipinfo==4.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
PyJWT==2.5.0
pyparsing==3.0.9
redis==4.3.4
requests==2.28.1
rq==1.11.1
SQLAlchemy==1.4.41
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
yarl==1.8.1
zipp==3.8.1
